% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{EmbedPlot}
\alias{EmbedPlot}
\title{Embedding plot for sections}
\usage{
EmbedPlot(object, group = "section", method = "TSNE")
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{group}{Character, a variable in slot \code{meta} defining the groups of sections. Sections in the same group have same colors.}

\item{method}{Character, the embeddings for scatter plot. Must be one of \code{"TSNE"}, \code{"UMAP"}, or \code{"PCA"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Scatter plot for sections with two-dimenstional embeddings in a \code{SummarizedExperiment} object. Each point stands for a section.
}
\examples{
data(zh.data)
zh <- CreateTomo(zh.data)
zh <- Normalize(zh)
zh <- TSNE(zh)
# Plot TSNE embeddings.
EmbedPlot(zh)

# Plot UMAP embeddings.
zh <- UMAP(zh)
EmbedPlot(zh, method="UMAP")

# Color sections by kmeans cluster labels.
zh <- KMeans(zh, 3)
EmbedPlot(zh, group="kmeans_cluster")
}
