% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{GetMatrix}
\alias{GetMatrix}
\title{Get a matrix from \code{SummarizedExperiment} object}
\usage{
GetMatrix(object, matrix)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{matrix}{Character, must be one of \code{"count"}, \code{"normalized"}, or \code{"scaled"}.}
}
\value{
A numeric matrix. Returns \code{NA} if such matrix does not exist.
}
\description{
Get one of three stored matrics, \code{count}, \code{normalized}, or \code{scaled}, in a \code{SummarizedExperiment} object.
}
\examples{
data(zh.data)
zh <- CreateTomo(zh.data)

# Get the raw read count matrix.
matrix_count <- GetMatrix(zh, "count")

# Get the normalized read count matrix.
zh <- Normalize(zh)
matrix_normalized <- GetMatrix(zh, "normalized")

# Get the scaled read count matrix.
zh <- Scale(zh)
matrix_scaled <- GetMatrix(zh, "scaled")
}
