% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{Normalize}
\alias{Normalize}
\title{Normalize data}
\usage{
Normalize(object)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}
}
\value{
A \code{SummarizedExperiment} object with normalized read count matrix saved in assay \code{'normalized'}.
}
\description{
Normalize the raw read count in a \code{SummarizedExperiment} object.
}
\details{
This function should be run for \code{SummarizedExperiment} object created from raw read count matrix.
If the \code{SummarizedExperiment} object already has a normalized count matrix. The function simply return the original object.
Library sizes of all sections are normalized to the median library size.
}
\examples{
data(zh.data)
zh <- CreateTomo(zh.data)
zh <- Normalize(zh)
}
