% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{Scale}
\alias{Scale}
\title{Scale data}
\usage{
Scale(object)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}
}
\value{
A \code{SummarizedExperiment} object with scaled read count matrix saved in assay \code{'scaled'}.
}
\description{
Scale the normalized read count in a \code{SummarizedExperiment} object.
}
\details{
This function should be run for \code{SummarizedExperiment} object with normalized read count matrix.
The normalized read counts of each gene are subjected to Z score transformation across sections.
}
\examples{
data(zh.data)
zh <- CreateTomo(zh.data)
zh <- Normalize(zh)
zh <- Scale(zh)
}
