% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zh.data}
\alias{zh.data}
\title{A raw read count matrix of zebrafish injured heart.}
\format{
A numeric matrix with 16495 genes as rows and 40 section as columns. Its row names are gene names and column names are section names.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE74652}
}
\usage{
data(zh.data)
}
\description{
A dataset containing gene expression across 40 sections of \strong{z}ebrafish \strong{h}eart generated with Tomo-seq.
The zebrafish heart is 3 days post cryoinjury (3 dpi).
}
\references{
Wu CC, Kruse F, Vasudevarao MD, et al. Spatially Resolved Genome-wide Transcriptional Profiling Identifies BMP Signaling as Essential Regulator of Zebrafish Cardiomyocyte Regeneration. \emph{Dev Cell}. 2016;36(1):36-49. doi:10.1016/j.devcel.2015.12.010
}
\keyword{datasets}
