% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateTomo}
\alias{CreateTomo}
\title{Create a \code{SummarizedExperiment} object}
\usage{
CreateTomo(matrix.count = NULL, matrix.normalized = NULL, min.section = 3)
}
\arguments{
\item{matrix.count}{A numeric matrix or matrix-like data stucture that can be coverted to matrix, with genes with rows, sections as columns and values as raw read counts.
Columns should be sorted according to section numbers.}

\item{matrix.normalized}{A numeric matrix or matrix-like data stucture that can be coverted to matrix, with genes with rows, sections as columns and values as normalized read counts.
Columns should be sorted according to section numbers.}

\item{min.section}{An integer. Genes expressed in less than \code{min.section} sections will be filtered out.}
}
\value{
A \code{SummarizedExperiment} object
}
\description{
Create a \code{SummarizedExperiment} object from raw read counts or normalized read counts.
}
\details{
This is the method to create a \code{Tomo} object, which is an instance of class \code{SummarizedExperiment}. At least one of raw read count matrix and normalized raw count matrix (like FPKM, TPM, ...) must be used for input.
}
\examples{
data(zh.data)
zh <- CreateTomo(zh.data)
}
